Generator in development
Created a new package to hold the main program, added a main function,
developed a quick testingTxt function for automatically testing the
setup so far.

modified Generator to be public, along with setting up the run()
function within.

Made markovThreeState an abstract, and set it up to automatically create
an ArrayList<String>.

Made progress on the Pattern Analyzer. It now saves all known sentences
and parses them, saving the unique words, and ordering them.

Modified PostGenerator to be public as well, and fixed a simple bug.

made ReadTwitter public, and ReadTxtFile public.