Major update
Changed from a hashmap to a hashtable for the markovChain.

Changed the sentence headers and footers to __STRT and __END - easier to
identify.

Removed the timer function from Generator - no need in the class.

Cleaned junk text from MarkovThreeState.

Changed the Sentence destructor in PatternAnalyzer so it can handle the
Vector hashtable.

Removed processInputData from hashTable, as not needed.

Finished the prototype generate function in Post Generator.

Removed the toggleUpdateSave - no need to toggle in here.

Cleaned ReadTxtFile.

Added documentation to testReadSystem.

Added documentation to all files.