Local prototype finished.
Well, it now functions. Upside? it works. Downsides:

Repeat sentences.

Repeated phrases.

Spouts gibberish.

Some of this can be solved with more data (the tantarus file is fairly
lacking), but clearly more dimensionality is needed for sentences to
make sanity. Will be upgrading from a two-state setup to a three state -
should mitigate that issue.