Progress!
After a while of being gone, due to a class, I am starting work on this
project again! Huzzah!

Slight changes to the general plan - going to go ahead and modify the
system so it is more intelligent-ish. Adding Some data mining to the
system, going to modify the markov chain to utilize Retweet and Favorite
data to attempt to create tweets that will be shared more.

Finished twitter integration on the download approach, have the system
saving tweet info (Just Text, Number of Retweets and Number of likes) to
a local file. Also Modified the ReadTwitter class to have a
ArrayList<Status> system so I can pass the status data (with the
favorites and retweet data) to the PatternAnalyzer.

googleStatusUpdates is an example of data pulled, as of today.