Developing main UI/Reader capabilities
Probably broke my Markov chain. Whoops.

Developed a text UI, with a main menu system and loop for running.

Created a relatively automated way to handle connecting to twitter,
currently testing out in case 0, a test case for development, and likely
the example case for when it fully runs.

Created a config.txt file, that will be used to automate connecting to
twitter.

Aiming to finish implementing up to the post-processor, which will be
started soon.