Good and Bad news.
Good news - fixed a bug.

Bad news - PA constructor needs to be rewritten from a while to a do
loop to avoid " " items being put into a sentence. Then after that fix,
need to fix the bug within markov creation helper thingy that divides it
up, because it's thinking the " " is a thing I want to keep.

Destroy with prejudice.