Fixed duplicate data issue
Since I'm storing weights in the hashtable, I don't need duplicates
within the markov chain - Thus created markovChainContains to use as a
boolean check.

- added markovChainContains
- created checks to reduce items stored within the markovChain.
- modified debugging code so one can see the words, their weights, etc
- renammed markovChain OBJ in PatternAnalyzer - was getting confusing.
Need to refactor and rename stuff.
- could potentially optimize PA code.

Probably ready to move on to the generator, or go back and
optimize/clean code, and set up the other choices within the menu.